change project class name etc
