--- brave-spark-framework-filter/src/main/java/com/github/kristofa/brave/spark/SparkHttpServerRequest.java
@@ -1,36 +1,0 @@
-package com.github.kristofa.brave.spark;
-
-import com.github.kristofa.brave.http.HttpServerRequest;
-import spark.Request;
-
-import javax.servlet.http.HttpServletRequest;
-import java.net.URI;
-
-
-class SparkHttpServerRequest implements HttpServerRequest {
-    private Request request;
-
-    public SparkHttpServerRequest(Request request) {
-        this.request = request;
-    }
-
-    @Override
-    public String getHttpHeaderValue(String headerName) {
-        return request.headers(headerName);
-    }
-
-    @Override
-    public URI getUri() {
-        HttpServletRequest httpServletRequest = request.raw();
-        StringBuffer url = httpServletRequest.getRequestURL();
-        if (httpServletRequest.getQueryString() != null && !httpServletRequest.getQueryString().isEmpty()) {
-            url.append('?').append(httpServletRequest.getQueryString());
-        }
-        return URI.create(url.toString());
-    }
-
-    @Override
-    public String getHttpMethod() {
-        return request.requestMethod();
-    }
-}
