--- brave-spark-framework-filter/src/main/java/com/github/kristofa/brave/spark/SparkHttpServerResponse.java
@@ -1,18 +1,0 @@
-package com.github.kristofa.brave.spark;
-
-import com.github.kristofa.brave.http.HttpResponse;
-import spark.Response;
-
-class SparkHttpServerResponse implements HttpResponse {
-
-    private Response response;
-
-    public SparkHttpServerResponse(Response response) {
-        this.response = response;
-    }
-
-    @Override
-    public int getHttpStatusCode() {
-        return response.raw().getStatus();
-    }
-}
