+++ brave-sparkjava/src/test/resources/log4j2.properties
